feat: overhaul theme to allow RGB colours and border styles
